mod 2 tunnin muistiinpanot mod2tunti.py:ssa.
02 tehtävät tehty 02tehtävät.py -tiedostoon.

21.8.2025 klo 0:49, muutin kaksi komentoa (rivit 55, 56) floatista intiksi, syystä että kyseisissä kohdissa on tarkoitus syöttää kokonaislukuja ja float on tässä tapauksessa tarpeeton, toimii kyllä mutta int sopii tässä yhteydessä paremmin.

+++ mod02/02tehtävät.py
@@ -30,9 +30,9 @@ suorakulmion_piiri = print(f'suorakulmion_piiri_on: {2 * suorakulmion_kanta + 2 
 # Tehtävä 4 Kirjoita ohjelma, joka kysyy kolme kokonaislukua.
         # Ohjelma tulostaa lukujen summan, tulon ja keskiarvon.
 
-kokonaisluku1 = float(input('Mikä on ensimmäinen kokonaisluku?'))
-kokonaisluku2 = float(input('Mikä on toinen kokonaisluku?'))
-kokonaisluku3 = float(input('Mikä on kolmas kokonaisluku?'))
+kokonaisluku1 = int(input('Mikä on ensimmäinen kokonaisluku?'))
+kokonaisluku2 = int(input('Mikä on toinen kokonaisluku?'))
+kokonaisluku3 = int(input('Mikä on kolmas kokonaisluku?'))
 
 lukujen_summa = print(f'lukujen summa: {kokonaisluku1 + kokonaisluku2 + kokonaisluku3}')
 
@@ -52,8 +52,8 @@ lukujen_keskiarvo = print(f'lukujen keskiarvo: {(kokonaisluku1 + kokonaisluku2 +
 
 #kysytään arvot keskiaikaisina mittoina
 
-leiviskat = float(input('Anna leivisköjen määrä:'))
-naulat = float(input('Anna naulojen määrä:'))
+leiviskat = int(input('Anna leivisköjen määrä:'))
+naulat = int(input('Anna naulojen määrä:'))
 luodit = float(input('Anna luotien määrä:'))
 
 # Muunnetaan kaikki saadut arvot grammoiksi
